OBPIH-5811 Move function for displaying status to StockMovement domain
